requirements.txt
requests==2.23
pydantic==1.5.1
PyHamcrest==2.0.2
PyYAML==5.3.1